% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs_var_tf.R
\name{bs_var_tf}
\alias{bs_var_tf}
\title{Estimate variance of a trendfiltering fit by bootstrapping}
\usage{
bs_var_tf(Y, lambda, sample.size, READS = NULL, sds = NULL,
  calc.sds = TRUE, positions = NULL, scale.pos = NULL, ord = 2,
  n.rep = 100)
}
\arguments{
\item{Y}{A matrix of complete (not average) data for one group.}

\item{READS}{A matrix of reads corresponding to \code{Y} (Optional). If \code{READS} is provided,
the elements in \code{Y} are assumed to be binomial observations and \code{sds} will be
calculated unless \code{calc.sds=FALSE}}

\item{sds}{A matrix of estimated standard deviations of \code{Y}. If missing
observations are assumed to have equal variance.}

\item{scale.pos}{}

\item{ord}{Order of trendfiltering. Can be 0, 1 or 2.}

\item{n.rep}{Number of bootstrap samples to take.}

\item{pos}{Vector of positions corresponding to the rows of \code{Y}}
}
\value{
A list with three elements
\describe{
  \item{\code{all.fits}}{All of the fits from bootstrapped samples. p by n.rep matrx}
  \item{\code{avg.fit}}{Equivalent to \code{rowMeans(all.fits)}}
  \item{\code{var}}{Estimated variance.}
}
}
\description{
Estimate variance of a trendfiltering fit by bootstrapping
}

