% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation_utils.R
\name{collect_methylation_data}
\alias{collect_methylation_data}
\title{Take multiple data frames with data from a single condition and
make data ready for JADE}
\usage{
collect_methylation_data(dfs)
}
\arguments{
\item{dfs}{A list of data frames. Each data frame
should have three columns - position, reads, and counts}
}
\value{
A list with elements counts, reads, pos, probs, and sds.
}
\description{
Take multiple data frames with data from a single condition and
make data ready for JADE
}

