% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation_utils.R
\name{combine_replicates}
\alias{combine_replicates}
\title{Combine data across replicates}
\usage{
combine_replicates(dfs, range = NULL)
}
\arguments{
\item{dfs}{A list of data frames. Each data frame
should have three columns - position, reads, and counts}
}
\value{
A combined data frame with three columns - position (pos), number of reads (reads), and
number of counts (y)
}
\description{
Combine data across replicates
}

