% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv_err_sds.R
\name{cv_err_wts}
\alias{cv_err_wts}
\title{Use cross validation paths to choose a value of gamma}
\usage{
cv_err_wts(orig.path, cv.path.list = NULL, use.converged.only = TRUE,
  control.l1 = TRUE)
}
\arguments{
\item{orig.path}{Either a path object or a file containing a \code{jade_path} object.}

\item{cv.path.list}{Either a list of \code{jade_path} objects or a vector of file names
containing \code{jade_path} objects for the cross validation data sets.}

\item{use.converged.only}{Only use fits which have converged.}

\item{control.l1}{Only use fits with \code{l1.total <= l1.total0}.}
}
\value{
A list with elements
#' \describe{
 \item{\code{l1.total}}{A vector of length N where N is the number of fits in \code{orig.path} giving the total
 L1 distance between all pairs of profiles. Equivalent to \code{orig.path$l1.total}.}
 \item{\code{cv.err.l1}}{An \code{n.folds} by N matrix where N is the number of fits in \code{orig.path}.
 Each row gives the average corss validation error for each value of \code{l1.total}.}
 \item{\code{err.l1}}{A vector of length N giving average cross validation error over all folds.
 Equivalent to \code{colSums(cv.err.l1)} with missing values for fits that were discarded according to
 specification of \code{use.converged.only} and \code{control.l1}.}
 \item{\code{err.se.l1}}{Estimated standard error of \code{err.l1}.}
 \item{\code{cv.min.l1,cv.1se.l1}}{Indices of the fits in \code{orig.path} corresponding to the minimum and 1
 standard error rule cross validation error.}
}
}
\description{
This function requires having a \code{jade_path} object for the original data
and for data with missing folds. See \code{\link{cv_fit0}}.
}

