% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv_fit0.R
\name{cv_fit0}
\alias{cv_fit0}
\title{Run initial cross validation fits for JADE}
\usage{
cv_fit0(orig.fit, n.folds = 5, which.fold = 1:n.folds, data.file = NULL,
  n.rep.bs.var = NULL, save.prefix = NULL, return.objects = TRUE,
  lambda = NULL)
}
\arguments{
\item{orig.fit}{JADE fit using all the data or the name of a file containig a JADE fit.}

\item{n.folds}{Number of cross validation folds.}

\item{which.fold}{Vector of folds to fit. If This can be used to run cv fits in parallel.
If \code{which.fold=1:n.folds} or is left missing cv fits will be run sequentially.}

\item{data.file}{Optional file containing data. This is required if you want to bootstrap the variance.}

\item{n.rep.bs.var}{Number of reps to use in bootstrapping variance of the fit. If NULL varaince will
not be estimated.}

\item{save.prefix}{If provided each fit will be saved to file named save.prefix.fold.RData}

\item{return.objects}{If \code{TRUE} a list of jade objects will be returned. Otherwise
nothing will be returned.}

\item{lambda}{If \code{NULL} \eqn{\lambda} will be chosen by cross validation for each fold. Otherwise
\code{lambda} should be a list of length \code{which.fold} giving \eqn{lambda} for each fold.}
}
\value{
A single jade object or list of jade objects of the same type as
\code{orig.fit} (either \code{jade_admm} or \code{jade_gd}) if \code{return.objects=TRUE}
}
\description{
Use an existing JADE fit at \eqn{\gamma = 0} to start cross validation fits with the same parameters.
These can be used as start computation of a series of solutions for \eqn{\gamma} between 0 and \eqn{\gamma_{max}}{\gamma_max}
using either \code{\link{jade_path_plan}} or \code{\link{jade_path}}.
}

