% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jade_util.R
\name{get_regions_total}
\alias{get_regions_total}
\title{Calculate total number of separated regions for each fit in a path}
\usage{
get_regions_total(path, tol)
}
\arguments{
\item{path}{Path of JADE fits}

\item{tol}{Tolerance for determining separation}
}
\value{
A vector of same length as path$gammas
}
\description{
Calculate total number of separated regions for each fit in a path
}

