% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jade_util.R
\name{get_sep_total}
\alias{get_sep_total}
\title{Recalculate sep.total for a path that has already been run}
\usage{
get_sep_total(path, tol)
}
\arguments{
\item{path}{Path of JADE fits}

\item{tol}{Tolerance for determining separation}
}
\value{
A vector of same length as path$gammas
}
\description{
Recalculate sep.total for a path that has already been run
}

