% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_separated_regions.R
\name{get_separated_regions}
\alias{get_separated_regions}
\title{Get a table of separated regions.}
\usage{
get_separated_regions(fit, which.window = 1, chr = "chr22",
  new.tol = NULL, data.range = NULL)
}
\arguments{
\item{fit}{A JADE fit.}

\item{which.window,chr}{Will be put in the table. Can be useful for writing other files.}

\item{new.tol}{Recalculate separation using a different tolerance.}
}
\value{
A list of two tables of separated regions.
One of the tables is merged over partition types, the other is not.
Returns 0 if all profiles are fused or only. These tables ca be passed
to the \code{\link{plot_jade}} function in \code{sep.tab} argument.
separated at singleton sites.
}
\description{
Get a table of separated regions.
}

