% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jade_admm.R
\name{jade_admm}
\alias{jade_admm}
\title{Fit JADE}
\usage{
jade_admm(y, gamma, pos = NULL, scale.pos = NULL, lambda = NULL,
  sample.size = NULL, ord = 2, sds = NULL, fit.var = NULL,
  var.wts = NULL, subset.wts = NULL, theta0 = NULL, u.alpha0 = NULL,
  u.beta0 = NULL, verbose = FALSE, tol = 0.001, max.it = 1000,
  rho.alpha = NULL, rho.beta = 1, adjust.rho.alpha = FALSE,
  tau.decr = 2, tau.incr = 2, mu = 10, e.rel = 1e-04, e.abs = 1e-08)
}
\arguments{
\item{y}{Data matrix of size p x K. May contain NA values but may not contain rows which are all NA.}

\item{gamma}{Fusion penalty.}

\item{pos}{Position vector of length p. If missing will use 1:p.}

\item{scale.pos}{An integer indicating to internally scale positions to range between 0 and \code{scale.pos}.}

\item{lambda}{Smoothing penalty vecor of length K.
If not provided, lambda will be chosen by cross validation.}

\item{sample.size}{Vector of sample sizes of length K.
If missing sample sizes are assumed to be 1.}

\item{ord}{Order of polynomial to fit. May be 0, 1, or 2.}

\item{sds}{Matrix of estimated standard deviations of size p x K.
These are the inverse of the diagonal elements of \eqn{A){i}}{A_i}.
Only the relative sizes of \code{sds} is important.}

\item{fit.var}{Matrix of size p x K of estimated variance of trendfiltering fits.
This will be used to construct the pairwise weight matrices \eqn{W}{W}.
Currently this is only supported for \eqn{K=2}.
\code{fit.var} can be estimated by bootstrapping. See \code{\link{bs_var}}.}

\item{var.wts}{If \code{fit.var} is not provided, the diagonal
elements of \eqn{W} may be specified here.
Since pairwise weights are currently only allowed for \eqn{K=2},
\code{var.wts} must be a vector of length p.}

\item{subset.wts}{This option can be used to obtain a de-biased fit with the
\eqn{\gamma} penalty only applied to pairs of points previously determined to be fused.
It should be a list of lists of the same format as the output of \code{\link{get_sep}}.
Elements are vectors of length p of 0s and 1s with 0 indicating
that the pair of points should not be penalized.}

\item{theta0,u.alpha0,u.beta0}{Starting values for \eqn{\theta} and the dual variables.
If a solution has been found for a nearby value
of \eqn{\gamma} using these values can improve convergence time.
If not provided the solution at \eqn{\gamma = 0} is used.}

\item{verbose}{Be chattier.}

\item{tol}{Tolerance for declaring points separated.
Separation can be recalculated with a different value of \code{tol} using \code{\link{get_sep}}.}

\item{max.it}{Maximum number of iterations.}

\item{rho.alpha,rho.beta}{ADMM step sizes.
\code{rho.alpha} has length K and \code{rho.beta} is a constant.
Change with caution.}

\item{adjust.rho.alpha}{Adaptively change rho.alpha.
This does not seem to help so the default is FALSE.}

\item{tau.incr,tau.decr,mu}{Parameters for adjusting step size. Change with caution.}

\item{e.rel,e.abs}{Parameters for determining convergence. Change with caution.}
}
\value{
A \code{jade_tf} object. This really just a list with values including
\describe{
  \item{\code{fits}}{A p x K matrix of solutions.}
  \item{\code{n}}{Number of iterations to convergence}
  \item{\code{beta},\code{alpha}}{See paper.}
  \item{\code{u.alpha}{u.beta}}{Dual variables. See paper.}
  \item{\code{sep}}{List of lists giving separation. See \code{\link{get_sep}}}
}
As well as all of the original parameters.
}
\description{
This is the main function in the jadeTF package. It fits JADE by minimizing
\deqn{\sum_{i=1}^{K} \Big( \frac{n_{i}}{2}\Vert A_{i}(y_{i} - \theta_{i}) \Vert_{2}^{2}  +
\lambda_{i}\Vert D^{k+1, x}\theta_{i}\Vert_{1}\Big) +
\gamma \sum_{i < l}\Vert W_{il} (\theta_{i} - \theta_{l}) \Vert_{1}}{
\sum_i=1^K ( n_i/2 || A_i y_i - \theta_i ||_2^2  +
\lambda_i || D \theta_i||_1) +
\gamma \sum_{j < l}|| W_jl (\theta_j - \theta_l) ||_1}
}

