% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path_tf.R
\name{jade_path_guess}
\alias{jade_path_guess}
\title{Fit JADE at a sequence of gamma values without knowing much about where to start.}
\usage{
jade_path_guess(fit0, n.fits, out.file, temp.file = NULL, max.it = 10000,
  log.gamma.min = -3, log.gamma.max = 20, start.step = 0.03,
  tol = 0.001, max.fits = 10 * n.fits, buffer = 0.01,
  restart.file = NULL, verbose = TRUE)
}
\arguments{
\item{fit0}{Either a jade object or a file containing a jade object produced by fitting with \code{gamma=0}}

\item{n.fits}{Desired number of fits along the path. Actual results may vary.}

\item{out.file}{Name of a file to save the results to.}

\item{temp.file}{Name a temp file. If missing will use a default based on \code{out.file}.
This file will be deleted at the end.}

\item{max.it}{Maximum number of jade iterations.}

\item{max.fits}{Maximum number of fits.}

\item{restart.file}{Provide if restarting from a temporary file.}

\item{log.gamma.minlog.gamma.max}{Smallest and largest values of \code{gamma} that are allowed.}

\item{hard.stop}{Stop at \code{log.gamma.stop} regardless of if the path is done or not.}
}
\value{
The output of this function is both returned and saved to a file. Partial results
are saved to a temporary file along the way. The path objects is a list with several components:
\describe{
 \item{\code{gammas}}{A list of length n of gamma values at which JADE was fit. The first element
 is always zero.}
  \item{\code{JADE_fits}}{A list of n JADE objects in the same order as \code{gammas}.}
  \item{\code{l1.total}}{Vector of length n giving the total L1 distance between pairs of profiles at each value of \code{gamma}}
  \item{\code{sep.total}}{Vector of length n giving the total number of separated sites between all pairs of profiles}
  \item{\code{sep}}{List of lists of matrices giving the pairwise separation
  between profiles. sep[[i]][[j-i]] is a \eqn{p \times n} matrix which describes the
  separation between profiles for group i and group j.}
  \item{\code{tol}}{The tolerance at which the seaparation in \code{sep.total} and \code{sep}
  were calculated}
}
}
\description{
This function requires having previously fit the data at \code{gamma}=0. JADE will be fit
at a serries of \code{gamma} values The function tries to fit JADE at a series of \code{gamma}
values so that fits occur at evenly spaced values of \code{l1.total}.
}

