% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jade_permute.R
\name{jade_permute}
\alias{jade_permute}
\title{Run JADE with permuted group assignments}
\usage{
jade_permute(Y, fit0, gammas, save.prefix, tol = 0.001, which.perm = 1:100,
  n.rep.fit.var = NULL, sd.type = c("Orig", "Equal", "Binomial", "Poisson"),
  READS = NULL)
}
\arguments{
\item{Y}{Full data matrix (p by n)}

\item{fit0}{A JADE fit using original group labels at gamma=0}

\item{gammas}{List of gammas at which to fit permutation fits}

\item{tol}{Tolerance for declaring sites separated}

\item{n.rep.fit.var}{If NULL use original fit.var. Otherwise
the number of bootstrap replications to use.}

\item{sd.type}{If "Orig" use sds used in fit0. Otherwsie recalculated sds
for each permutation assuming binomial or poisson models.}

\item{READS}{Required if sd.type="Binomial"}

\item{n.perm}{Number of permutations}
}
\value{
A list with elements
#' \describe{
 \item{\code{sep.total}}{A matrix length(gammas) by h n.perm }
 \item{\code{gammas}}{Gamma values fit for each permutation}
  \item{\code{tol}}{The tolerance at which the seaparation in \code{sep.total} and \code{sep}
  were calculated}
}
}
\description{
This function takes the complete data matrix and re-runs
JADE permuting the group labels. It can optionally recalculate the weights
in the likelihood.
}

