% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylation_utils.R
\name{methylation_windows}
\alias{methylation_windows}
\title{Find window boundaries for methylation data}
\usage{
methylation_windows(positions, gapsize = 2000, mindata = 20, nmiss = NULL,
  nm.edges = FALSE)
}
\arguments{
\item{positions}{Vector of positions}

\item{gapsize}{Maximum gap to allow between data points in a window}

\item{mindata}{Minimum number of data points per window}

\item{nmiss}{If nm.edges=TRUE provide a vector listing the number
of missing points at each position.}

\item{nm.edges}{Make sure the last data point in each window is non-missing in all samples}
}
\value{
A list with elements counts, reads, pos, probs, and sds.
}
\description{
Find window boundaries for methylation data
}

