% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_jade.R
\name{plot_jade}
\alias{plot_jade}
\title{Plot a JADE fit}
\usage{
plot_jade(fits, pos, y = NULL, reads = NULL, cols = NULL, range = NULL,
  wsize = 10, maxcov = Inf, take.log.cov = FALSE, maxwidth = 3,
  minwidth = 0.5, ylim = NULL, sep.tab = NULL, ylab = NULL,
  xlab = "Position", bg.color = "blue")
}
\arguments{
\item{fits}{A p by K matrix of fitted values (the \code{fits}) element of a JADE fit object.}

\item{pos}{Vector of positions.}

\item{y}{A p by K matrix of data. Optional. If provided, data points will be added to the plot.
Data can be plotted by itself using the \code{plot_data} function.}

\item{reads}{A p by K matrix of read counts or an indication of the relative amount of data for
each element of \code{y}. Line widths will be proportional to \code{reads}.}

\item{wsize}{Window size used to determine line widths.}

\item{sep.tab}{A table giving regions which are separated.
This can be obtained using \code{\link{get_separated_regions}}.}
}
\value{
A \code{ggplot} object for the desred plot.
}
\description{
Plot a JADE fit
}

