% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jade_util.R
\name{sep_gd}
\alias{sep_gd}
\title{Determine which pairs of sites are separated from a jade_gd object.}
\usage{
sep_gd(obj, tol = 1e-06)
}
\arguments{
\item{obj}{A jade_gd object
for a \code{jade_admm} object use \code{fits=obj$beta}.}

\item{tol}{Tolerance for determining separation}
}
\value{
A list of lists of length p vectors containing 0s and 1s.
The vector stored in \code{[[i]][[j-i]]} indicates the separation between group i and group j.
}
\description{
Determine which pairs of sites are separated from a jade_gd object.
}

