% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tf_weights.R
\name{trendfilter_weights}
\alias{trendfilter_weights}
\title{Solve trendfiltering problem with weights.}
\usage{
trendfilter_weights(y, pos, wts = NULL, ord = 0, approx = FALSE,
  maxsteps = 2000, minlam = 0, rtol = 1e-07, btol = 1e-07,
  eps = 1e-04, verbose = FALSE)
}
\arguments{
\item{y}{Vector of data.}

\item{pos}{Vector of positions.}

\item{wts}{Vector of weights}

\item{ord}{Order of polynomial to fit.}

\item{...}{See \code{\link[genlasso]{trendfilter}} for other options.}
}
\value{
Returns a \code{genlasso} object with additional elements \code{weights} with the weights provided.
}
\description{
This function minimizes
\deqn{\frac{1}{2} \Vert w*(y - \theta) \Vert^{2}_{2} + \lambda\Vert D\theta\Vert_{1}}{
1/2 || w(y - \theta) ||^2 + \lambda_||D\theta||_1}
Where D is a trendfiltering pendalty matrix.
This function borrows code heavily from the \code{genlasso} package and
also returns a \code{genlasso} object.
}

