% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path_planned.R
\name{jade_path_planned}
\alias{jade_path_planned}
\title{Fit JADE at a sequence of gamma values.}
\usage{
jade_path_planned(fit0, out.file, log.gamma.start, n.fits = NULL,
  log.gamma.stop = NULL, step.size = NULL, tol = 0.001, verbose = TRUE,
  temp.file = NULL, max.it = 10000, log.gamma.max = 20,
  restart.file = NULL, hard.stop = FALSE)
}
\arguments{
\item{fit0}{Either a jade object or a file containing a jade object produced by fitting with \code{gamma=0}}

\item{out.file}{Name of a file to save the results to.}

\item{log.gamma.start}{Starting value of \code{log10(gamma)}.}

\item{n.fits}{Desired number of fits along the path. Actual results may vary.
\code{n.fits} can only be missing if \code{step.size} is provided.}

\item{log.gamma.stop}{Speculative stopping point. If \code{log.gamma.stop} and \code{n.fits}
are provided \code{step.size = (log.gamma.stop-log.gamma.start)/n.fits}.}

\item{step.size}{Alternative to specifying \code{log.gamma.stop} and \code{n.fits}}

\item{temp.file}{Name a temp file. If missing will use a default based on \code{out.file}.
This file will be deleted at the end.}

\item{max.it}{Maximum number of jade iterations.}

\item{log.gamma.max}{Absolute largest value of \code{gamma} that will be fit.}

\item{restart.file}{Provide if restarting from a temporary file.}

\item{hard.stop}{Stop at \code{log.gamma.stop} regardless of if the path is done or not.}
}
\value{
The output of this function is both returned and saved to a file. Partial results
are saved to a temporary file along the way. The path objects is a list with several components:
\describe{
 \item{\code{gammas}}{A list of length n of gamma values at which JADE was fit. The first element
 is always zero.}
  \item{\code{JADE_fits}}{A list of n JADE objects in the same order as \code{gammas}.}
  \item{\code{l1.total}}{Vector of length n giving the total L1 distance between pairs of profiles at each value of \code{gamma}}
  \item{\code{sep.total}}{Vector of length n giving the total number of separated sites between all pairs of profiles}
  \item{\code{sep}}{List of lists of matrices giving the pairwise separation
  between profiles. sep[[i]][[j-i]] is a \eqn{p \times n} matrix which describes the
  separation between profiles for group i and group j.}
  \item{\code{tol}}{The tolerance at which the seaparation in \code{sep.total} and \code{sep}
  were calculated}
}
}
\description{
This function requires having previously fit the data at \code{gamma}=0. JADE will be fit
at a serries of \code{gamma} values evenly spaced on the log scale. The spacing is determined by
a starting point, a stopping point and the number of fits desired.

The start and end points should be pretty good guesses but don't need to be exact.
If \code{log.gamma.start} is too large \code{jade_path_planned} will back up until it finds a value of
\code{gamma} with no fusion between profiles. Similarly if \code{log.gamma.stop} is too small and the path isn't complete,
the function will keep going until all the profiles are fused.\code{log.gamma.stop}.
Guessing the start and stop locations can have a few negative consequences:

-If \code{log.gamma.start} is much too large it may take a (prohibitively) long time
to fit the initial fit or convergence might not be reached.
Warm starts help a lot with convergence times so starting in the middle of the path can be costly.

- If \code{log.gamma.start} and \code{log.gamma.stop} are way too far apart, the chosen step size will be too large
which will have a similar effect as starting in the middle of the path. It will also result in a sparsely populated
path which might not be very useful.

- If the endpoints are way too close together the step size will be
way small and many more fits than \code{n.fits} will be calculated.
}

